+++ Data/include/Camps1te/Data.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "Data/DataFile.h"
+#include "Data/DataStore.h"
+#include "Data/JsonDataFile.h"
+#include "Data/JsonDataStore.h"
