+++ Data/include/Camps1te/Data/DataStore.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace Camps1te::Data {
+
+    class DataStore {
+    public:
+    };
+}
