+++ Data/include/Camps1te/Data/JsonDataStore.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "DataStore.h"
+
+namespace Camps1te::Data {
+
+    class JsonDataStore {
+    public:
+    };
+}
