+++ Data/include/Camps1te/Data/JsonKeys.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <_Log_.h>
+
+#include <string>
+#include <vector>
+
+namespace Camps1te::Data::JsonKeys {
+
+    constexpr auto TOPLEVEL_PLUGINS = "plugins";
+}
