--- Editor/include/Camps1te/Editor/MiscExamples/TEMPLATE.h
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <QLabel>
-#include <QPushButton>
-#include <QVBoxLayout>
-#include <QWidget>
-
-
-namespace Camps1te::Editor::MiscExamples::ImageWindow {
-
-    class Instance {
-    public:
-        Instance() = default;
-    };
-}
