+++ MapView/xmake.lua
@@ -1,0 +1,5 @@
+target("Camps1te.MapView")
+    set_kind("headeronly")
+    add_includedirs("include", {public = true})
+    add_headerfiles("include/(**).h")
+    add_packages("_Log_", "spdlog", "string_format")
