+++ UI/include/Camps1te/UI/msgbox.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define msgbox(...) QMessageBox::information(nullptr, "", string_format(__VA_ARGS__).c_str())
