+++ UI/xmake.lua
@@ -1,0 +1,4 @@
+target("Camps1te.UI")
+    set_kind("headeronly")
+    add_includedirs("include", {public = true})
+    add_headerfiles("include/(**).h")
